requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-decouple>=3.8
mysqlclient>=2.1.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0